Event Sets

Event Sets are user-defined groups of events (preset or native). The user is free to allocate and use any number of them provided the required resources can be provided.

Multiple event sets can be used simultaneously and can even share counter values.

Adding and removing events

Events can be added to event set using push! and append!. Removal of events is done using empty! and delete!

Base.push! — Method

push!(evtset::EventSet, evt::Event)

Adds one PAPI event to the set. The event can be either a native or preset event.

Base.append! — Method

append!(evtset::EventSet, evt::Event)

Adds multiple PAPI events to the set. The events can be either a native or preset events.

Base.empty! — Method

empty!(evtset::EventSet)

Remove all PAPI events from the set.

Base.delete! — Method

delete!(evtset::EventSet, evt::Event)

Remove a PAPI events from the set. The event should have previously been added using push! or append!.