Counting Events

PAPI provides the ability to start, stop, read and accumulate the counters for a specified list of events. PAPI.jl exposes this functionality (See counting), but builds more user-friendly primitives on top of this.

PAPI.jl stores count values using the Counts type.

Measuring Performance

PAPI,jl provides two ways of measuring performance counters:

Quick, one shot profile

Extensive sample based

Profile

profile and @profile can be used to quickly measure some performance counts on a function or expression. Since the function is only executed once, these counts should be taken with a grain of salt as they can be riddled with noise especially for short running functions.

PAPI.profile — Function

profile(f::Function, events::Vector{Event}; gcfirst::Bool=true, warmup::Int64=0)

Execute the function f once while counting specific events.

Arguments:

-`f`: the function to profile
-`events`: the events to count
-`gcfirst`: run the gc several times before the execution to reduce gc noise
-`warmup`: number of times to run the function prior to counting

Return values:

EventValues containing the events, counts and runtime collected

PAPI.@profile — Macro

profile(ex, args...)

Convience macro for profiling an expression. Events can be specified as a first argument, otherwise the default events [BR_INS, BR_MSP, TOT_INS, TOT_CYC] are counted

Arguments and return values are similar to profile

Example:

@profile f(x, y, z) # sampling default events
@profile [PAPI.TOT_INS, PAPI.DP_OPS, native_event] f(x, y, z) gcfirst=false

The resulting EventValues contain the events and counts collected and when printed a nice summary is generated with additional information.

Sample

Contrary to profile, sample and @sample perform multiple executions. The resulting samples can be used to investigate the distributions of the counts and to perform various statistical tests.

Note

You should be careful when making decisions on performance based on single or aggregate values (such as mean, minimum, median) as well as "eyeball" statistics. Statistical tests such as a t-test, wilcoxon can be used.

Personally, I'm also wary towards any asymptotical tests putting strong assumptions on the distributions. For example, performance numbers (even runtime) are not always normal distributed: the process is usually skewed towards one side. I tend to use the samples themselves as well as their quantiles.

PAPI.sample — Function

sample(f::Function, events::Vector{Event}; max_secs::Float64=5, max_epochs::Int64=1000, gcsample::Bool=false, warmup::Int64=1)

Execute the function f several times, each time counting specific events. Sampling continues until either the maximum number of samples max_epochs are collected or the runtime budget max_secs is exceeded.

Arguments:

-`f`: the function to profile
-`events`: the events to count
-`max_secs`: maximum number of seconds to sample for
-`gcsample`: run the gc several times before the execution to reduce gc noise
-`warmup`: number of times to run the function prior to counting

Return values:

EventStats containing the events, counts and runtime collected

PAPI.@sample — Macro

sample(ex, args...)

Convience macro for sampling an expression. Events can be specified as a first argument, otherwise the default events [BR_INS, BR_MSP, TOT_INS, TOT_CYC] are counted

Arguments and return values are similar to sample

Example:

@sample f(x, y, z) # sampling default events
@sample [PAPI.TOT_INS, PAPI.DP_OPS, native_event] f(x, y, z) max_secs=1

The resulting EventStats contain the events and all counts collected and when printed a nice summary is generated with additional statistics. The additional statistics are computed using the mean and are mostly intended as information to guide the optimization process.

Example

As an example consider the following function

function mysum(X::AbstractArray)
    s = zero(eltype(X))
    for x in X
        s += x
    end
    s
end

Next, we'll create an array of double precision values and produce a profile using the default events

Note

warmup is set to 1, to force compilation prior to measurement

X = rand(10000)
@profile mysum(X) warmup=1

EventValues:
  BR_INS = 10283 # 25.0% of all instructions # 644.0 M/sec 
  BR_MSP = 36 # 0.0%[m of all branches
  TOT_INS = 41516 # 0.63 insn per cycle
  TOT_CYC = 65906 # 4.128 Ghz # 1.587 cycles per insn
  runtime = 15965 nsecs

Since the number of elements added was 10000, you can note it take roughly 4 instructions per element. Out of which one is a branch instruction, corresponding to the termination test of the loop. We also expect the data needs to be loaded in every iteration, something we can verify by additionally measuring the number of loads using the preset PAPI.LD_INS.

@profile [PAPI.TOT_INS, PAPI.TOT_CYC, PAPI.LD_INS, PAPI.BR_INS] mysum(X)

EventValues:
  TOT_INS = 41023 # 0.751 insn per cycle
  TOT_CYC = 54607 # 3.396 Ghz # 1.331 cycles per insn
  LD_INS = 10328 # 25.0% of all instructions # 642.0 M/sec 
  BR_INS = 10212 # 25.0% of all instructions # 635.0 M/sec 
  runtime = 16080 nsecs

Indeed, in addition to the branch instruction there is a load in every iteration as well. Generally, you would expect a load, a branch, an addition and a counter increment in every iteration, and this data seems to confirm this. However modern processors are capable of loading and multiplying multiple data items at the same time. The @simd macro might be useful here.

Note

@simd instructs the compiler that it is safe to execute the iterations in arbitrary and overlapping order. This also means that we are fine with any floating-point errors introduced by this reordering of operations.

function mysum2(X::AbstractArray)
    s = zero(eltype(X))
    @simd for x in X
        s += x
    end
    s
end

@profile [PAPI.TOT_INS, PAPI.TOT_CYC, PAPI.LD_INS, PAPI.BR_INS] mysum2(X) warmup=1

EventValues:
  TOT_INS = 5416 # 0.133 insn per cycle
  TOT_CYC = 40794 # 1.133 Ghz # 7.532 cycles per insn
  LD_INS = 2828 # 52.0% of all instructions # 79.0 M/sec 
  BR_INS = 839 # 15.0% of all instructions # 23.0 M/sec 
  runtime = 36015 nsecs

Note the reduction in the number of cycles, instructions, branches, loads and runtime. From these measurements, one could assume the addition of @simd has a positive effect. However, it would be better if we compute more samples and perform a statistical test to confirm that the effect is significant.

Finally, let's look at the number of floating-point operations and vectorized floating-point operations.

@profile [PAPI.TOT_INS, PAPI.DP_OPS, PAPI.VEC_DP] mysum(X)

EventValues:
  TOT_INS = 41029
  DP_OPS = 10000 # 24.0% of all instructions # 1.0x vectorized
  VEC_DP = 10000 # 24.0% of all instructions
  runtime = 15121 nsecs

@profile [PAPI.TOT_INS, PAPI.DP_OPS, PAPI.VEC_DP] mysum2(X)

EventValues:
  TOT_INS = 5422
  DP_OPS = 10015 # 185.0% of all instructions # 3.998x vectorized
  VEC_DP = 2505 # 46.0% of all instructions
  runtime = 5863 nsecs

Both functions perform the same number of additions ($\approx 1$ per element), but mysum is able to vectorize those operations (performing almost 4 additions at the same time). Looking back at the earlier comparison, we can also see the number of loads has decreased by a factor of 4, indicating that the loop now loads roughly four elements simultaneously. The number of branches had decreased even more. Investigation of @code_native mysum2(X) indeed also indicated that the compiler does loop unrolling in addition to vectorization.

Let's now collect multiple samples in order to draw our final conclusion:

stats = @sample mysum(X) # original version
stats2 = @sample mysum2(X) # optimized version

EventStats:
  BR_INS = [814, 814, 814, 814, 814, 814, 814, 814, 814, 814  …  814, 814, 814, 814, 814, 814, 814, 814, 814, 814] # 16.0% of all instructions # 772.0 M/sec 
  BR_MSP = [21, 4, 2, 2, 2, 2, 2, 2, 2, 2  …  2, 2, 2, 2, 2, 2, 2, 2, 2, 2] # 0.0%[m of all branches
  TOT_INS = [5168, 5168, 5168, 5168, 5168, 5168, 5168, 5168, 5168, 5168  …  5168, 5168, 5168, 5168, 5168, 5168, 5168, 5168, 5168, 5168] # 1.289 insn per cycle
  TOT_CYC = [8137, 4175, 3948, 4068, 3925, 4174, 3996, 3967, 3983, 3968  …  3993, 4017, 3993, 3976, 3991, 3993, 3944, 3949, 3962, 4057] # 3.802 Ghz # 0.778 cycles per insn
  runtime = [2202, 1162, 1111, 1145, 1105, 1177, 1131, 1121, 1122, 1120  …  1003, 1011, 1004, 1001, 1003, 1008, 989, 994, 995, 1026] nsecs

1000 samples were collected for both version for the default events as well as the runtimes. A decision can for example be made using a wilcoxon (or mann-whitney U) test. This tests whether the two sets of samples come from the same underlying distribution.

using HypothesisTests
x = stats.time
y = stats2.time
MannWhitneyUTest(x,y)

Approximate Mann-Whitney U test
-------------------------------
Population details:
    parameter of interest:   Location parameter (pseudomedian)
    value under h_0:         0
    point estimate:          10466.0

Test summary:
    outcome with 95% confidence: reject h_0
    two-sided p-value:           <1e-99

Details:
    number of observations in each group: [1000, 1000]
    Mann-Whitney-U statistic:             9.99034e5
    rank sums:                            [1.49953e6, 5.01466e5]
    adjustment for ties:                  335820.0
    normal approximation (μ, σ):          (4.99034e5, 12912.9)

Rejection of the null-hypothesis indicates the underlying distributions are different and the changes have some significant effect (in this case a positive effect).

Note

Alternatively, a student t-test could be used to perform a similar test. However, this test assumes normally distributed samples. Plotting a histogram of the samples clearly indicates this is not the case.

Using a Monte-Carlo estimator, the expected speedup can be computed as well as 95% uncertainty intervals.

using Statistics
println("expected speedup = ", mean(x ./ y'))
println("95% of the time, the speedup lies within ", quantile(vec(x ./ y'), [.025, 0.975]))

expected speedup = 11.359966780769762
95% of the time, the speedup lies within [10.010462074978204, 13.27035175879397]

PAPI interface

Only a number of events can be counted at the same time, num_counters returns the number of hardware counters available on the system.

PAPI.num_counters — Function

Get the number of hardware counters available on the system

PAPI.num_counters() initializes the PAPI library if necessary.

PAPI_num_counters()` returns the number of hardware counters available on the system.

After selecting a number of events to count, the counting process can be started and stopped using start_counters and stop_counters.

Warning

It is the user’s responsibility to choose events that can be counted simultaneously by reading the vendor’s documentation.

PAPI.start_counters — Function

start_counters(events)

Start counting hardware events. This function cannot be called if the counters have already been started.

PAPI.stop_counters — Function

stop_counters(evts::Vector{Event})

Stop counters and returns counts The counters must have been started by a previous call to start_counters

PAPI.stop_counters! — Function

stop_counters!(values::Vector{Counts})

Stop counters and return current counts. The counters must have been started by a previous call to start_counters

The user must provide a vector of the correct size (equal to the number of events)

events = Event[PAPI.TOT_INS, PAPI.DP_OPS]
start_counters(events)
computation()
values = stop_counters(events)

2-element Array{Int64,1}:
 57837298
    10001

During the counting process, the current counts can be queried and accumulated using read_counters! and accum_counters!.

PAPI.read_counters! — Function

read_counters!(values::Vector{Counts})

Read and reset counters. read_counters! copies the event counters into values. The counters are reset and left running after the call.

The user must provide a vector of the correct size (equal to the number of events)

PAPI.accum_counters! — Function

accum_counters!(values::Vector{Counts})

Accumulate and reset counters. accum_counters! accumulates the event counters into values. The counters are reset and left running after the call.

The user must provide a vector of the correct size (equal to the number of events)

events = Event[PAPI.DP_OPS]
values = Vector{Counts}(undef, length(events))
start_counters(events)
computation(100) # perform 100 double precision operations
read_counters!(values)
println(values[1], " roughly equals 100")

computation(100) # perform 100 double precision operations
accum_counters!(values)
println(values[1], " roughly equals 200")

values[1] = -100
computation(100) # perform 100 double precision operations
accum_counters!(values)
println(values[1], " roughly equals 0")
stop_counters(events); nothing

100 roughly equals 100
200 roughly equals 200
0 roughly equals 0