Public Documentation

Documentation for PAPI.jl's public interface.

See the Internals section of the manual for internal package docs covering all submodules.

Contents

Public Documentation

Contents

Index

Public Interface

Index

PAPI.accum_counters!

PAPI.num_counters

PAPI.profile

PAPI.read_counters!

PAPI.sample

PAPI.start_counters

PAPI.stop_counters

PAPI.stop_counters!

Public Interface

PAPI.accum_counters! — Method

accum_counters!(values::Vector{Counts})

Accumulate and reset counters. accum_counters! accumulates the event counters into values. The counters are reset and left running after the call.

The user must provide a vector of the correct size (equal to the number of events)

PAPI.num_counters — Method

Get the number of hardware counters available on the system

PAPI.num_counters() initializes the PAPI library if necessary.

PAPI_num_counters()` returns the number of hardware counters available on the system.

PAPI.profile — Method

profile(f::Function, events::Vector{Event}; gcfirst::Bool=true, warmup::Int64=0)

Execute the function f once while counting specific events.

Arguments:

-`f`: the function to profile
-`events`: the events to count
-`gcfirst`: run the gc several times before the execution to reduce gc noise
-`warmup`: number of times to run the function prior to counting

Return values:

EventValues containing the events, counts and runtime collected

PAPI.read_counters! — Method

read_counters!(values::Vector{Counts})

Read and reset counters. read_counters! copies the event counters into values. The counters are reset and left running after the call.

The user must provide a vector of the correct size (equal to the number of events)

PAPI.sample — Method

sample(f::Function, events::Vector{Event}; max_secs::Float64=5, max_epochs::Int64=1000, gcsample::Bool=false, warmup::Int64=1)

Execute the function f several times, each time counting specific events. Sampling continues until either the maximum number of samples max_epochs are collected or the runtime budget max_secs is exceeded.

Arguments:

-`f`: the function to profile
-`events`: the events to count
-`max_secs`: maximum number of seconds to sample for
-`gcsample`: run the gc several times before the execution to reduce gc noise
-`warmup`: number of times to run the function prior to counting

Return values:

EventStats containing the events, counts and runtime collected

PAPI.start_counters — Method

start_counters(events)

Start counting hardware events. This function cannot be called if the counters have already been started.

PAPI.stop_counters! — Method

stop_counters!(values::Vector{Counts})

Stop counters and return current counts. The counters must have been started by a previous call to start_counters

The user must provide a vector of the correct size (equal to the number of events)

PAPI.stop_counters — Method

stop_counters(evts::Vector{Event})

Stop counters and returns counts The counters must have been started by a previous call to start_counters